refactor(indicators): simplify emaRibbon code
