+++ documentation/indicators/moving-averages.md
@@ -30,6 +30,6 @@ SMMA is similar to EMA but applies a longer smoothing factor, resulting in a slo
 
 This smoothing method is based on the technique introduced by J. Welles Wilder and is often used in indicators like ATR and RSI. It is a form of exponential smoothing that reduces lag while maintaining stability.
 
-## Super Guppy
+## EMA Ribbon
 
-Super Guppy combines 22 EMAs with increasing periods to form two ribbons (fast and slow). It tracks the alignment and spread of these EMAs to assess trend direction and strength. It’s useful for identifying sustained trends and spotting early reversals.
+EMA Ribbon stacks a series of EMAs with progressively longer periods. By default it uses 22 EMAs starting at period 3 and increasing by 3 each time. The indicator tracks the spread between the highest and lowest EMA values to gauge trend strength.
