+++ src/indicators/movingAverages/emaRibbon/emaRibbon.test.ts
@@ -1,0 +1,62 @@
+import { describe, expect, it } from 'vitest';
+import { EMARibbon } from './emaRibbon.indicator';
+
+const ribbon = new EMARibbon({ count: 22, start: 1, step: 1 });
+describe('EMARibbon', () => {
+  it.each`
+    candle                                                                                      | expected
+    ${{ close: 81, open: 81, high: 82.96289647361662, low: 79.03710352638338, volume: 403 }}    | ${null}
+    ${{ close: 24, open: 81, high: 83.85720988022568, low: 21.142790119774318, volume: 814 }}   | ${null}
+    ${{ close: 75, open: 24, high: 76.94326596315126, low: 22.056734036848734, volume: 1064 }}  | ${null}
+    ${{ close: 21, open: 75, high: 79.67167346434113, low: 16.328326535658874, volume: 330 }}   | ${null}
+    ${{ close: 34, open: 21, high: 34.711649023641215, low: 20.28835097635878, volume: 964 }}   | ${null}
+    ${{ close: 25, open: 34, high: 36.18138133787512, low: 22.818618662124877, volume: 214 }}   | ${null}
+    ${{ close: 72, open: 25, high: 73.33035016836122, low: 23.669649831638775, volume: 860 }}   | ${null}
+    ${{ close: 92, open: 72, high: 94.97523624952838, low: 69.02476375047162, volume: 486 }}    | ${null}
+    ${{ close: 99, open: 92, high: 101.5127586628106, low: 89.4872413371894, volume: 647 }}     | ${null}
+    ${{ close: 2, open: 99, high: 99.0804764241746, low: 1.9195235758253941, volume: 396 }}     | ${null}
+    ${{ close: 86, open: 2, high: 86.08306699694582, low: 1.916933003054178, volume: 252 }}     | ${null}
+    ${{ close: 80, open: 86, high: 87.6552826540483, low: 78.3447173459517, volume: 299 }}      | ${null}
+    ${{ close: 76, open: 80, high: 80.75068906338092, low: 75.24931093661908, volume: 469 }}    | ${null}
+    ${{ close: 8, open: 76, high: 77.2154050332975, low: 6.784594966702496, volume: 605 }}      | ${null}
+    ${{ close: 87, open: 8, high: 90.4594244098795, low: 4.540575590120497, volume: 261 }}      | ${null}
+    ${{ close: 75, open: 87, high: 91.6843769949034, low: 70.3156230050966, volume: 926 }}      | ${null}
+    ${{ close: 32, open: 75, high: 77.42509022428217, low: 29.57490977571783, volume: 544 }}    | ${null}
+    ${{ close: 65, open: 32, high: 66.96254158787707, low: 30.037458412122934, volume: 819 }}   | ${null}
+    ${{ close: 41, open: 65, high: 69.41431179251751, low: 36.58568820748249, volume: 1012 }}   | ${null}
+    ${{ close: 9, open: 41, high: 43.081607806555546, low: 6.918392193444454, volume: 736 }}    | ${null}
+    ${{ close: 13, open: 9, high: 15.19184083967761, low: 6.808159160322391, volume: 971 }}     | ${null}
+    ${{ close: 26, open: 13, high: 28.498919258834377, low: 10.501080741165623, volume: 1098 }} | ${{ results: [26, 22.623787541861265, 23.55510139465332, 25.881656538393806, 28.466690567829, 30.937937372603297, 33.116368234689745, 34.923156284844175, 36.359736979000886, 38.20405673168798, 39.12982913986841, 39.915578295973496, 40.60739056114612, 42.52146966410521, 42.55529502232869, 42.8306960225572, 44.302223577029245, 44.62635679932202, 45.74505263157895, 47.96417233560091, 49.85281385281385, 51.04545454545455], spread: 28.42166700359328 }}
+    ${{ close: 56, open: 26, high: 59.7249059948739, low: 22.2750940051261, volume: 379 }}      | ${{ results: [56, 44.87459584728709, 39.77755069732666, 37.92899392303628, 37.644460378552665, 38.09852669471664, 38.83727617601731, 39.60689933265658, 40.28778958320071, 41.43968278047198, 41.94152428322368, 42.39010471197757, 42.80633476669667, 44.318607042224514, 44.23588314453761, 44.38002590225636, 45.60197651291489, 45.82358239939338, 46.770547368421056, 48.72948925601987, 50.4116489571035, 51.476284584980235], spread: 18.355539621447335 }}
+    ${{ close: 28, open: 56, high: 58.44863123776439, low: 25.551368762235615, volume: 117 }}   | ${{ results: [28, 33.62486528242903, 33.88877534866333, 33.95739635382177, 34.42964025236844, 35.21323335336903, 36.12795713201298, 37.027588369844004, 37.83023166656057, 38.99610409311344, 39.617936902686395, 40.176242448596405, 40.69114408574001, 42.142792769927915, 42.206397751470405, 42.452964031402665, 43.646201344813235, 43.94741583103618, 44.893492631578944, 46.755252184017976, 48.37422632463955, 49.43486853411239], spread: 21.43486853411239 }}
+    ${{ close: 65, open: 28, high: 66.775319062114, low: 26.224680937886014, volume: 1041 }}    | ${{ results: [65, 54.541621760809676, 49.444387674331665, 46.37443781229306, 44.619760168245634, 43.72373810954931, 43.34596784900974, 43.24367984321201, 43.264185333248456, 43.72408516709282, 43.848280752238665, 43.99528207188927, 44.16383778777715, 45.19042040060419, 45.05559803253661, 45.10555649829647, 46.018845639833984, 46.16347732250606, 46.90414336842105, 48.4928472141115, 49.88566029512686, 50.788358226798266], spread: 21.756320156787993 }}
+    ${{ close: 58, open: 65, high: 68.65939501847261, low: 54.340604981527385, volume: 510 }}   | ${{ results: [58, 56.84720725360322, 53.72219383716583, 51.02466268737584, 49.07984011216375, 47.802670078249506, 47.0094758867573, 46.522862100276, 46.21134826659876, 46.319706045803215, 46.206900626865554, 46.149854060829384, 46.14043238952327, 46.8983643471903, 46.67364827846953, 46.62254985143806, 47.35008501318576, 47.40942707803174, 48.01372903157895, 49.39829033657708, 50.623327541024416, 51.415457511424506], spread: 11.85956761047673 }}
+    ${{ close: 17, open: 58, high: 60.01918882495998, low: 14.980811175040019, volume: 878 }}   | ${{ results: [17, 30.282402417867743, 35.361096918582916, 37.4147976124255, 38.386560074775836, 39.00190719874965, 39.50710691506798, 39.96222607799245, 40.36907861327901, 40.98885040111172, 41.3390838557213, 41.66526112839409, 41.97751347673423, 42.91191576756493, 42.96444224366084, 43.137543986562996, 43.977853345054015, 44.208434754028396, 44.91235612842105, 46.312738875950686, 47.56666140093129, 48.4228090321702], spread: 31.4228090321702 }}
+    ${{ close: 90, open: 17, high: 93.37632122668938, low: 13.623678773310617, volume: 150 }}   | ${{ results: [90, 70.09413413928925, 62.68054845929146, 58.4488785674553, 55.591040049850555, 53.572790856249746, 52.13033018630098, 51.081731393994126, 50.295262890623206, 49.8999685100005, 49.44923654643441, 49.10137480094885, 48.837868694343626, 49.19032699855627, 48.84388696320323, 48.65077410579088, 49.09142519560357, 49.02859951676225, 49.42112051557895, 50.47343041157443, 51.42423763721026, 52.03821694241627], spread: 41.34922589420912 }}
+    ${{ close: 87, open: 90, high: 91.79876207396401, low: 85.20123792603599, volume: 914 }}    | ${{ results: [87, 81.36471137976308, 74.84027422964573, 69.86932714047317, 66.06069336656704, 63.12342204017839, 60.84774763972574, 59.06356886199543, 57.63621031249857, 56.6454287809095, 55.70769712202868, 54.931932523879794, 54.28960173800883, 54.2316167320821, 53.61340109280283, 53.162447740403714, 53.30348906275873, 53.025589041313594, 53.17900846402105, 53.952151324757814, 54.65839785200933, 55.078371990901815], spread: 33.974410958686406 }}
+    ${{ close: 86, open: 87, high: 90.67917429835809, low: 82.32082570164191, volume: 582 }}    | ${{ results: [86, 84.45490379325436, 80.42013711482286, 76.32159628428391, 72.70712891104469, 69.659587171556, 67.1358107297943, 65.04944244821867, 63.308968249998856, 61.98262354801686, 60.75641426835723, 59.71163521251367, 58.81965863257899, 58.46740116780449, 57.661725956202474, 57.02568918270916, 56.9364347224522, 56.49657966854374, 56.46110761761895, 57.004327389066596, 57.507634410917575, 57.767209209084264], spread: 29.53889238238105 }}
+    ${{ close: 99, open: 86, high: 103.90666224637769, low: 81.09333775362231, volume: 804 }}   | ${{ results: [99, 94.15163459775145, 89.71006855741143, 85.39295777057035, 81.4714192740298, 78.04256226539714, 75.10185804734573, 72.59401079305897, 70.44717459999909, 68.71305563019561, 67.13034522363102, 65.75599902597311, 64.55970739935343, 63.87174767876389, 62.829010211677165, 61.96384339650808, 61.61016419773529, 60.97062391396019, 60.71499685585705, 61.00391525677454, 61.2796676462887, 61.35266927785955], spread: 38.28500314414295 }}
+    ${{ close: 3, open: 99, high: 100.63598126886475, low: 1.3640187311352432, volume: 436 }}   | ${{ results: [3, 33.38387819925048, 46.355034278705716, 52.435774662342205, 55.314279516019866, 56.601830189569384, 57.0763935355093, 57.12867506126808, 56.957739679999264, 56.76522733379641, 56.44195435302586, 56.101229945054165, 55.765463485160076, 55.7555146549287, 55.35038393521752, 55.02692064397772, 55.09792373132026, 54.86845297564859, 54.94349717027135, 55.47973285136744, 55.98151604208064, 56.27852412326306], spread: 54.12867506126808 }}
+    ${{ close: 70, open: 3, high: 74.01339408473842, low: -1.0133940847384295, volume: 747 }}   | ${{ results: [70, 57.794626066416825, 58.17751713935286, 59.46146479740533, 60.20951967734658, 60.429878706835275, 60.30729515163197, 59.9889694920974, 59.56619174399941, 59.17154963674252, 58.70162862752154, 58.23950226119968, 57.79896870156578, 57.654779367604874, 57.18158594331533, 56.78845939174505, 56.75370998339579, 56.46124739926453, 56.449147453244215, 56.862615436951494, 57.255923674618764, 57.47169593863149], spread: 13.550852546755785 }}
+    ${{ close: 1, open: 70, high: 70.62974231530183, low: 0.37025768469818, volume: 722 }}      | ${{ results: [1, 19.93154202213894, 29.58875856967643, 36.076878878443196, 40.47301311823105, 43.4499133620252, 45.48047136372398, 46.880309604964644, 47.85295339519953, 48.59490424824388, 49.08469052293462, 49.43342499024588, 49.68483031562781, 50.10080878525756, 50.15888770040092, 50.225111228010334, 50.55885331857403, 50.62322135723668, 50.90423270791979, 51.542366347718016, 52.141748795107965, 52.56111368309832], spread: 51.56111368309832 }}
+    ${{ close: 27, open: 1, high: 30.879348918826008, low: -2.879348918826009, volume: 221 }}   | ${{ results: [27, 24.64384734071298, 28.294379284838215, 32.446127327065916, 35.98200874548737, 38.749938115732284, 40.860353522792984, 42.46246302608361, 43.68236271615962, 44.668558021290444, 45.403908769112185, 45.98212883790036, 46.44414027053812, 47.02070094722322, 47.2640267378508, 47.49274520118559, 47.941202949843586, 48.13656647752756, 48.51380943712781, 49.20499812412583, 49.85613526827997, 50.3384081454376], spread: 25.694560804724617 }}
+    ${{ close: 9, open: 27, high: 29.207668220474442, low: 6.792331779525556, volume: 148 }}    | ${{ results: [9, 14.21461578023766, 18.647189642419107, 23.06767639623955, 26.988005830324912, 30.24995579695163, 32.895265142094736, 35.02636013139836, 36.7458901729277, 38.18336565378309, 39.33659064092682, 40.29257055514646, 41.09497737474696, 41.95127415426012, 42.48102339561945, 42.96418694222258, 43.61440262208318, 44.01692790094571, 44.562428493415034, 45.37595068373289, 46.14194115298179, 46.74376395887781], spread: 37.74376395887781 }}
+    ${{ close: 92, open: 9, high: 96.46225023362183, low: 4.5377497663781705, volume: 331 }}    | ${{ results: [92, 66.07153859341255, 55.323594821209554, 50.64060583774373, 48.65867055354994, 47.892825569251166, 47.671448856571054, 47.68716899108762, 47.79671213834216, 47.96820826218617, 48.113825534105686, 48.247559700508546, 48.367123464068825, 48.624437600358775, 48.67089547116702, 48.733106125490515, 48.99058010851839, 49.067777595583, 49.30618564407353, 49.81633633290119, 50.31085559361981, 50.67908883201886], spread: 44.328551143428946 }}
+    ${{ close: 68, open: 92, high: 94.82774764949542, low: 65.17225235050458, volume: 338 }}    | ${{ results: [68, 67.35717953113752, 61.66179741060478, 57.58436350264624, 55.10578036903329, 53.63773254946512, 52.75358664242829, 52.20113143751259, 51.83736971067373, 51.610352214515956, 51.42818794508807, 51.28639666966107, 51.171820112058995, 51.20784592031094, 51.08703353727114, 50.99979952249163, 51.102737874238564, 51.06064311183742, 51.17556707966617, 51.548113825005835, 51.918959630563464, 52.18525502053896], spread: 17.000200477508372 }}
+    ${{ close: 9, open: 68, high: 69.94866467256739, low: 7.051335327432617, volume: 823 }}     | ${{ results: [9, 28.45239317704584, 35.33089870530239, 38.150618101587746, 39.73718691268886, 40.88409467818937, 41.81518998182122, 42.600880006954235, 43.26989576853898, 43.863015448240326, 44.35682328757339, 44.78079718202091, 45.14727438176485, 45.58013313093615, 45.82615434511225, 46.058646637492615, 46.42465588821206, 46.6332069948019, 46.95801037169956, 47.49591250833861, 48.01723602778497, 48.430015453535574], spread: 39.430015453535574 }}
+  `('returns $expected when candle.close = $candle.close', ({ candle, expected }) => {
+    ribbon.onNewCandle(candle);
+    const result = ribbon.getResult();
+
+    if (expected === null) {
+      // Still warming-up: all EMAs not yet seeded
+      expect(result).toBeNull();
+    } else {
+      expect(result).not.toBeNull();
+      expected.results.forEach((value, idx) => {
+        expect(result!.results[idx]).toBeCloseTo(value, 13);
+      });
+      expect(result!.spread).toBeCloseTo(expected.spread, 13); // same 1e-13 tolerance as the TEMA test
+    }
+  });
+});
